investment_analyst/requirements.txt
openai-agents>=0.0.7
openai>=1.66.0
python-telegram-bot>=21.0
playwright>=1.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
yfinance>=0.2.36
aiosqlite>=0.20.0
matplotlib>=3.8.0
numpy>=1.26.0
